+++ DTCD-StyleSystem/src/stories/BaseExpander.stories.js
@@ -27,7 +27,7 @@ export default {
         '',
         'with_borderBottom',
         'theme_iconLeft',
-        'theme_rotate90',
+        'rotation_type2',
       ],
     },
     open: {